Establish Python virtual environment to run Django
Django is typically run in a Python virtual environment. All python
dependencies that a Django project requires are installed in the virtual
environment via pip (a package manager). The virtual environment is then
activated and used to run the Django server.

As our project is developed, we will need to install various
dependencies to the virtual environment. Managing the virtual
environment via SSH would quickly become complex. Since the virtual
environment is simply a directory, I have included the virtual
environment in the backend project repository, so that it can be managed
via git.